requirements.txt
duckdb==0.9.2
langchain==0.0.353
openai==1.6.1
pandas==2.1.4
streamlit==1.29.0
huggingface_hub==0.20.1
python-dotenv==1.0.0
plotly==5.18.0
pydantic==2.5.3
